// src/Types/Table.ts
export type TableType = {
  ANTIWKF: string; // Identificador
  ACUSWKF: string; // Código do Centro de custo
  ADCCWKF: string; // Descrição do centro de Custo
  ADNAWKF: string; // Descrição da natureza
  AHISWKF: string; // Histórico
  ADFIWKF: string; // Nome do Fornecedor
  AFILWKF: string; // Filial
  ANFOWKF: string; // Company name
  AFORWKF: string; // Fornecedor
  ALOJWKF: string; // Loja
  ANATWKF: string; // Natureza
  APARWKF: string; // Parcela
  APREWKF: string; // Prefixo
  ATIPWKF: string; // Tipo
  ATITWKF: string; // Numero do Título
  ALBLWKF: string; // Setor
  AVALWKF: string; // Valor
};